ErrorList lexer fixes
- Better handling of CMake status messages

With the current handling, lines starting with "-" are considered "Diff deleted" message
except for lines starting with "---" which are considered as "Diff message". A common
line is also the CMake status lines which generate lines starting with "--". This fix
handles them as "normal" lines and not has "Diff deleted" lines

- Exclude output of the command "ls -l" from labeled as "Diff deleted" lines

A common output for the "ls -l" command are lines with a similar pattern to:

`-rw-r--r--  1 user user   80864 Nov  4 23:35 file.cpp`

These lines are mistakenly considered as "Diff deleted" lines - this commit
fixes this behavior

Signed-off-by: Eran Ifrah <eran@codelite.org>